Simplified adding generators
Not as organized as where the code was initially going but much easier to use... just pull the experiment string from MOA and use.

I really need to simplify design. This may lead to software design that isn't fantastic but my objective here is to run experiments, not to write perfectly designed software.